// src/types/banner-json.ts
export type JsonMeasureUnit = 'px' | 'mm' | 'cm' | 'in'

export type JsonUrlTarget = '_blank' | '_self' | '_parent' | '_top'

export const JsonLayerTypeArr = [
  'embed',
  'image',
  'shape',
  'svg',
  'video',
  'audio',
  'text',
  'button',
  'youtube',
  'clipart',
  'countdown',
  'rain',
  'snow',
  'fog',
] as const

export const NotAcceptedLayer = ['menu']

export type JsonLayerType = (typeof JsonLayerTypeArr)[number]

export type JsonDesignBackground = JsonBackground & {
  borderColor?: string
  useBorder: boolean
}
// ...
export type JsonDesignFeed = EditorDesignFeed | AdStudioDesignFeed

export function isEditorDesignFeed(
  variable: JsonDesignFeed
): variable is EditorDesignFeed {
  return typeof variable === 'object' && variable !== null && 'hash' in variable
}

export function isAdStudioDesignFeed(
  variable: JsonDesignFeed
): variable is AdStudioDesignFeed {
  return (
    typeof variable === 'object' && variable !== null && 'hashes' in variable
  )
}

export type JsonDesignProperties = {
  sortIndex?: number
  name: string
  width: number
  height: number
  loop: boolean
  loopCount?: number
  imageQuality?: number
  bannerSize?: string
  customSize?: boolean
  presetSize?: boolean
  backgroundColor?: JsonDesignBackground
  lastId: number
  version: string
  showGuidelines: boolean
  measureUnit: JsonMeasureUnit
  sizeVariation?: string | null
  feed?: JsonDesignFeed
  fallbackImagePosition?: number
  fallbackImageSlideIndex?: number
  bannerUrl: string
  useAsClickTag: boolean
}
// ...
export type JsonSlideOrElement = JsonSlide | JsonElement
// ...
export type JsonSlide = {
  properties: JsonSlideProperties
  type: 'slide'
  elements: JsonElement[]
}
// ...
export type JsonElement =
  | JsonText
  | JsonButton
  | JsonImage
  | JsonSvg
  | JsonAudio
  | JsonVideo
  | JsonShape
  | JsonEmbed
  | JsonYoutube
  | JsonClipart
  | JsonCountdown
  | JsonSnow
  | JsonRain
  | JsonFog
// ...
export type JsonDesign = {
  properties: JsonDesignProperties
  elements: JsonSlideOrElement[]
  resources?: Record<string, TResource>
}

export type DesignJsonWithHash = {
  hash: string
  userId: number
  banner: JsonDesign
}
// ...
export type TResource = {
  content: string
  usages: number
}